Backend/requirements.txt
# Core Rasa components
rasa==3.6.15
rasa-sdk==3.6.2

# ML and compatibility packages
numpy==1.23.4
protobuf==4.23.3

# Gemini fallback support
google-generativeai==0.3.2

# Env variable support
python-dotenv==0.21.0
